refactor: combine ERC-20 and ERC-721 voting token implementations and turn adapter into minimal proxy (#65)
**Motivation:**

This is a significant refactor of the entire token voting module. It
simplifies the contracts and provides a powerful architecture that has
broad token support and pushes complexity to the edges. Users can call
the deploy function directly from their Llama instance with a config for
the action creator, caster, and adapter, The adapter has a standardized
interface but a generalized configuration similar to strategies. We have
some standard checks that the adapter, creator, and caster are setup
correctly

**Modifications:**

- Removed ERC20 and ERC721 creator and caster contracts and moved
everything into generalized creator and caster contracts
- Turned the adapter into a minimal proxy that is initialized at
deployment with a generalized config
- Adds checks during deployment that the adapter was initialized
correctly
- Adds a `LlamaTokenAdapterTimestamp` logic contract that can be used
for IVotes tokens that use timestamp based checkpointing

**Result:**

Closes #60 (note this issue is outdated)
Closes #61 (note this issue is outdated)
Closes #63