refactor: remove governor role from configuration and determine it automatically (#86)
**Motivation:**

Rather than requiring Token Governors to specify an immutable governance
role at deploy time, this PR enables the use of multiples roles.

**Modifications:**

For token holder action creation, creators can specify the role they
want to create actions with on behalf of token holders when calling
`createAction` on `LlamaTokenGovernor`.

For token holder casting, this PR defines a `_getGovernanceRole`
function that is used by both `castVote` and `castVeto` as well as
`submitApproval` and `submitDisapproval`. `_getGovernanceRole` provides
a deterministic algorithm for determining which role token holders will
cast an approval or disapproval with. It does this by first checking if
the `LlamaTokenGovernor` has a force role for the action's strategy. It
iterates through all assigned roles from 1 to the max assigned role.

The token governor will cast with the first match if it's policy has a
force role. If it doesn't then it will fallback to using the
(dis)approvall role defined by the action's strategy.

**Result:**

Addresses
https://github.com/spearbit-audits/review-llama-token-governor/issues/6